Added color schemes and Merged ViewModels
