+++ .idea/misc.xml
@@ -27,6 +27,13 @@
         <entry key="..\:/studyMaterial/Projects/DigitalDiary/app/src/main/res/layout/fragment_group_details.xml" value="0.2598958333333333" />
         <entry key="..\:/studyMaterial/Projects/DigitalDiary/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.2717948717948718" />
         <entry key="..\:/studyMaterial/Projects/DigitalDiary/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.2717948717948718" />
+        <entry key="..\:/studyMaterial/Projects/ManageExpenses/app/src/main/res/layout/activity_main.xml" value="0.2598958333333333" />
+        <entry key="..\:/studyMaterial/Projects/ManageExpenses/app/src/main/res/layout/fragment_entry.xml" value="0.2598958333333333" />
+        <entry key="..\:/studyMaterial/Projects/ManageExpenses/app/src/main/res/layout/fragment_entry_details.xml" value="0.2598958333333333" />
+        <entry key="..\:/studyMaterial/Projects/ManageExpenses/app/src/main/res/layout/fragment_entry_groups.xml" value="0.2598958333333333" />
+        <entry key="..\:/studyMaterial/Projects/ManageExpenses/app/src/main/res/layout/fragment_entry_list.xml" value="0.2598958333333333" />
+        <entry key="..\:/studyMaterial/Projects/ManageExpenses/app/src/main/res/layout/fragment_group.xml" value="0.2598958333333333" />
+        <entry key="..\:/studyMaterial/Projects/ManageExpenses/app/src/main/res/layout/fragment_group_details.xml" value="0.2598958333333333" />
         <entry key="..\:/studyMaterial/Projects/a3-journalapp-satyansh1999/app/src/main/res/layout/activity_main.xml" value="0.2916666666666667" />
         <entry key="..\:/studyMaterial/Projects/a3-journalapp-satyansh1999/app/src/main/res/layout/fragment_entry.xml" value="0.208955223880597" />
         <entry key="..\:/studyMaterial/Projects/a3-journalapp-satyansh1999/app/src/main/res/layout/fragment_entry_details.xml" value="0.264" />
