+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity"
-    android:background="#25283D">
+    android:background="@color/fragment_background">
 
   <androidx.fragment.app.FragmentContainerView
       android:id="@+id/nav_host_fragment"
