+++ app/src/main/res/layout/fragment_entry.xml
@@ -5,7 +5,7 @@
     android:layout_margin="4dp"
     android:background="@color/design_default_color_secondary"
     android:orientation="vertical"
-    android:backgroundTint="#EFD9CE"
+    android:backgroundTint="@color/fragment_primary"
     android:paddingHorizontal="10dp"
     android:paddingVertical="6dp">
 
@@ -15,7 +15,7 @@
       android:layout_height="wrap_content"
       android:padding="2dp"
       android:text="@string/title_label"
-      android:textColor="#25283D"
+      android:textColor="@color/fragment_text"
       android:letterSpacing="0.1"
       android:textAppearance="@style/TextAppearance.AppCompat.Large" />
 
@@ -31,7 +31,7 @@
         android:layout_weight="0.55"
         android:padding="2dp"
         android:text="@string/date"
-        android:textColor="#25283D"
+        android:textColor="@color/fragment_text"
         android:textAllCaps="true"
         android:textStyle="italic" />
 
@@ -41,7 +41,7 @@
         android:layout_height="wrap_content"
         android:layout_weight="0.2"
         android:text="@string/group"
-        android:textColor="#25283D"
+        android:textColor="@color/fragment_text"
         android:textStyle="italic"/>
 
     <TextView
@@ -50,7 +50,7 @@
         android:layout_height="wrap_content"
         android:layout_weight="0.2"
         android:text="@string/amount"
-        android:textColor="#25283D"
+        android:textColor="@color/fragment_text"
         android:textStyle="bold|italic" />
   </LinearLayout>
 </LinearLayout>
