+++ app/src/main/res/layout/fragment_entry_details.xml
@@ -5,7 +5,7 @@
     android:id="@+id/layout_entry_detail"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#EFD9CE">
+    android:background="@color/fragment_primary">
 
   <TextView
       android:id="@+id/record_entry"
@@ -13,6 +13,8 @@
       android:layout_height="wrap_content"
       android:layout_marginTop="60dp"
       android:text="@string/record_entry"
+      android:textColor="@color/fragment_text"
+      android:textColorHint="@color/fragment_text"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintHorizontal_bias="0.498"
       app:layout_constraintStart_toStartOf="parent"
@@ -31,22 +33,12 @@
       android:inputType="text"
       android:minHeight="48dp"
       android:textAlignment="center"
+      android:textColor="@color/fragment_text"
+      android:textColorHint="@color/fragment_text"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintHorizontal_bias="0.0"
       app:layout_constraintStart_toStartOf="parent"
       app:layout_constraintTop_toBottomOf="@+id/record_entry" />
-
-  <Button
-      android:id="@+id/btn_entry_date"
-      style="?android:buttonBarButtonStyle"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:layout_marginTop="12dp"
-      android:text="@string/date"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintHorizontal_bias="0.498"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@+id/edit_amount" />
 
   <EditText
       android:id="@+id/edit_amount"
@@ -59,10 +51,25 @@
       android:inputType="numberDecimal"
       android:minHeight="48dp"
       android:textAlignment="center"
+      android:textColor="@color/fragment_text"
+      android:textColorHint="@color/fragment_text"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintHorizontal_bias="0.0"
       app:layout_constraintStart_toStartOf="parent"
       app:layout_constraintTop_toBottomOf="@+id/edit_title" />
+
+  <Button
+      android:id="@+id/btn_entry_date"
+      style="?android:buttonBarButtonStyle"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:layout_marginTop="12dp"
+      android:text="@string/date"
+      android:textColor="@color/fragment_text"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintHorizontal_bias="0.498"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintTop_toBottomOf="@+id/edit_amount" />
 
   <Button
       android:id="@+id/btn_save"
@@ -70,6 +77,7 @@
       android:layout_height="wrap_content"
       android:layout_marginTop="12dp"
       android:text="@string/save_label"
+      android:textColor="@color/fragment_button_text"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintHorizontal_bias="0.498"
       app:layout_constraintStart_toStartOf="parent"
