+++ app/src/main/res/layout/fragment_group.xml
@@ -5,19 +5,19 @@
     android:layout_margin="4dp"
     android:background="@color/design_default_color_secondary"
     android:orientation="horizontal"
-    android:backgroundTint="#EFD9CE"
+    android:backgroundTint="@color/fragment_primary"
     android:paddingHorizontal="10dp"
     android:paddingVertical="6dp"
     android:gravity="center_vertical">
 
-    <Button
+    <ImageButton
         android:layout_width="30dp"
         android:layout_height="30dp"
         android:id="@+id/edit_group"
         android:background="@android:drawable/ic_menu_edit"
+        android:backgroundTint="@color/fragment_text"
         android:layout_marginStart="5dp"
         android:layout_marginEnd="10dp"
-        android:backgroundTint="#63535B"
         android:title="@string/edit"/>
 
     <TextView
@@ -27,6 +27,6 @@
         android:padding="5dp"
         android:text="@string/title_label"
         android:letterSpacing="0.08"
-        android:textColor="#25283D"
+        android:textColor="@color/fragment_text"
         android:textAppearance="@style/TextAppearance.AppCompat.Large" />
 </LinearLayout>
