+++ app/src/main/res/values-night/themes.xml
@@ -2,13 +2,13 @@
   <!-- Base application theme. -->
   <style name="Theme.JournalApp" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
     <!-- Primary brand color. -->
-    <item name="colorPrimary">@color/purple_200</item>
-    <item name="colorPrimaryVariant">@color/purple_700</item>
-    <item name="colorOnPrimary">@color/black</item>
+    <item name="colorPrimary">@color/fragment_background</item>
+    <item name="colorPrimaryVariant">@color/fragment_button</item>
+    <item name="colorOnPrimary">@color/fragment_primary</item>
     <!-- Secondary brand color. -->
-    <item name="colorSecondary">@color/teal_200</item>
-    <item name="colorSecondaryVariant">@color/teal_200</item>
-    <item name="colorOnSecondary">@color/black</item>
+    <item name="colorSecondary">@color/fragment_button</item>
+    <item name="colorSecondaryVariant">@color/fragment_button</item>
+    <item name="colorOnSecondary">@color/fragment_background</item>
     <!-- Status bar color. -->
     <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
     <!-- Customize your theme here. -->
