+++ app/src/main/res/values/themes.xml
@@ -2,15 +2,15 @@
   <!-- Base application theme. -->
   <style name="Theme.JournalApp" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
     <!-- Primary brand color. -->
-    <item name="colorPrimary">@color/fragment_text</item>
-    <item name="colorPrimaryVariant">@color/purple_700</item>
+    <item name="colorPrimary">@color/fragment_background</item>
+    <item name="colorPrimaryVariant">@color/fragment_button</item>
     <item name="colorOnPrimary">@color/fragment_primary</item>
     <!-- Secondary brand color. -->
     <item name="colorSecondary">@color/fragment_button</item>
-    <item name="colorSecondaryVariant">@color/teal_700</item>
-    <item name="colorOnSecondary">@color/fragment_text</item>
+    <item name="colorSecondaryVariant">@color/fragment_button</item>
+    <item name="colorOnSecondary">@color/fragment_background</item>
     <!-- Status bar color. -->
-    <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
+    <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimary</item>
     <!-- Customize your theme here. -->
   </style>
 </resources>
